Add basic implementation of the exclusions list
